# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: exercises
    description: Python exercises operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /exercises:
    get:
      operationId: getExercises
      tags: [exercises]
      summary: Get all exercises
      description: Returns list of Python exercises
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
  /exercises/{id}/check:
    post:
      operationId: checkAnswer
      tags: [exercises]
      summary: Check user's answer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerSubmission"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
  /progress:
    get:
      operationId: getProgress
      tags: [exercises]
      summary: Get user progress
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Progress data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
  /gamification/status:
    get:
      operationId: getGamificationStatus
      tags: [exercises]
      summary: Get user gamification status (XP and League)
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Gamification status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamificationStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Exercise:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, tryhard]
        type:
          type: string
          enum:
            [
              multiple_choice,
              multiple_select,
              fill_blank,
              predict_output,
              bug,
              order,
              code,
            ]
        question:
          type: string
        options:
          type: array
          items:
            type: string
        hint:
          type: string
        explanation:
          type: string
        orderIndex:
          type: integer
        coins:
          type: integer
        xp:
          type: integer
        theme:
          type: string
        pytoReaction:
          type: string
        lesson:
          type: string
        skills:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - topic
        - difficulty
        - type
        - question
        - explanation
        - orderIndex
    AnswerSubmission:
      type: object
      properties:
        answer:
          type: string
        userId:
          type: string
          minLength: 1
        doubleOrNothing:
          type: boolean
          default: false
      required:
        - answer
        - userId
    AnswerResult:
      type: object
      properties:
        correct:
          type: boolean
        feedback:
          type: string
        correctAnswer:
          type: string
        xpEarned:
          type: integer
          minimum: 0
        alreadyCompleted:
          type: boolean
      required:
        - correct
        - feedback
        - correctAnswer
        - xpEarned
        - alreadyCompleted
    Progress:
      type: object
      properties:
        totalExercises:
          type: integer
        completedExercises:
          type: integer
        correctAnswers:
          type: integer
        topicProgress:
          type: object
          additionalProperties:
            type: integer
        totalAttempts:
          type: integer
        reviewExerciseIds:
          type: array
          items:
            type: integer
      required:
        - totalExercises
        - completedExercises
        - correctAnswers
        - topicProgress
        - totalAttempts
        - reviewExerciseIds
    GamificationStatus:
      type: object
      properties:
        totalXp:
          type: integer
        weeklyXp:
          type: integer
        leagueTier:
          type: integer
        boardId:
          type: string
        currentLevel:
          type: integer
        xpInCurrentLevel:
          type: integer
        xpNeededForNextLevel:
          type: integer
        percentage:
          type: integer
      required:
        - totalXp
        - weeklyXp
        - leagueTier
        - boardId
        - currentLevel
        - xpInCurrentLevel
        - xpNeededForNextLevel
        - percentage